backend/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
amqp==5.2.0
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
arrow==1.3.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
blinker==1.7.0
cachelib==0.9.0
celery==5.3.6
celery-redbeat==2.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dataclasses-json==0.6.4
Flask==3.0.1
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
flower==2.0.1
frozenlist==1.4.1
gunicorn==21.2.0
h11==0.14.0
humanize==4.9.0
ics==0.7.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
kombu==5.3.5
langchain==0.1.8
langchain-community==0.0.21
langchain-core==0.1.26
langsmith==0.1.5
MarkupSafe==2.1.4
marshmallow==3.20.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
outcome==1.3.0.post0
packaging==23.2
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic_core==2.16.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.4
PyYAML==6.0.1
redis==5.0.2
requests==2.31.0
selenium==4.17.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.27
TatSu==5.11.3
tenacity==8.2.3
tornado==6.4
trio==0.24.0
trio-websocket==0.11.1
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.1
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.4
